perf(secrets): answer teardown ownership from one list instead of per secret (#1264)
Closes #1263.

`down` inspected every compose-named secret for its `podup.project`
label and then, separately, fetched `/secrets/json` for the orphan sweep
— a call that already carries every label. Each label was fetched twice:
once per secret, and once for all of them.

### Measured

`bench/scenarios/secrets` (six `file:` secrets), Podman 5.7.0, `down
-v`:

| | before | after |
|---|---|---|
| `GET /secrets/NAME/json` | 6 | **0** |
| `DELETE /secrets/NAME` | 6 | 6 |
| `GET /secrets/json` | 1 | 1 |
| total for `down -v` | 18 | **12** |

### The list is a superset, not just a shortcut

Every secret podup creates carries `podup.project=<proj>`, so the
labelled set already contains everything the two compose-driven loops
reached, plus the orphans they miss. Teardown is therefore one pass over
that set rather than two loops and a sweep.

The guard is unchanged in substance. `/secrets/json` returns the same
`Spec.Labels` the per-secret inspect read, so a same-named secret the
user created by hand is absent from the set and survives, exactly as
before. The check is answered from a batch response instead of N single
ones.

### The two things this had to get right

**A failed list is `None`, not an empty set.** Since the list *is* the
ownership check now, collapsing a failure into "nothing is ours" would
delete nothing and report a clean teardown — a silent no-op dressed as
success. It falls back to the per-secret guarded path instead: the same
requests as before, reached only when the cheap route is unavailable.
The orphan sweep is not possible without a list, which is also how it
behaved before.

**The window widens.** It used to sit between each inspect and its
delete; it now sits between one list and up to N deletes, so it is wider
for the last name in the batch. `down` is best-effort here either way —
a delete failure is logged, not fatal — and a secret that changed hands
inside that window would have been created by podup moments earlier.
Recording the trade rather than leaving it implicit, since #1263 raised
it as the reason this was not a free rearrangement.

`delete_listed_secret` is a separate function rather than a flag on
`delete_secret`, so a call site cannot accidentally skip a check it was
supposed to make: the unchecked one is only reachable from a name the
list vouched for.

### Verified where it counts

Both majors, SELinux enforcing, including the two cases this change is
actually able to break:

```
Podman 5.8.1  Fedora 44          Enforcing   ALL PASS
Podman 6.0.1  Fedora 45 rawhide  Enforcing   ALL PASS
```

- an **unlabelled** secret with a project-scoped name survives `down`
- a **labelled** secret that no compose key names is still swept
- plus the existing coverage: every `file:` and `content:` secret read
back from inside the container across a cold up, a re-up and a teardown

### The module had to be split

The first CI run failed `line-limit`: `secrets/mod.rs` reached **591
code lines** against the 500 hard limit. It had been holding three
separate jobs, so the second commit splits it along those seams rather
than by code-versus-tests, with each module carrying its own tests the
way `plan.rs` already did.

```
mod.rs     591 -> 143   module docs, build_native_secrets, collect_payload_union
create.rs         238   create_project_secrets, create_secret
remove.rs         243   remove_internal_secrets and the teardown helpers
```

No behaviour change — the same 2007 tests pass. One thing worth flagging
for review: `pub(super)` meant `crate::engine` from `mod.rs` and means
`secrets` from a submodule, so both entry points are now spelled `pub(in
crate::engine)` explicitly rather than silently narrowing.

Full suite green (2007 tests), `cargo fmt --check` and `cargo clippy
--locked --all-targets --all-features -- -D warnings` clean.

### On the mutation testing

Four mutations, each turning the intended test red: collapsing a failed
list into an empty set, dropping the label filter, re-inspecting each
secret before deleting, and skipping orphans.

Worth recording how the first attempt went, because the failure mode is
indistinguishable from a useless test: my initial mutation of the
failed-list branch turned nothing red, and the reason was that rustfmt
had split the line it targeted, so the substitution never applied. A
mutation that does not reach the code looks exactly like an assertion
that cannot fail. Re-applied against the real line, it turns the
fallback test red as it should.

---------
